c8505/a2/requirements.txt

Pillow == 5.3.0
hashlib
Crypto